validator/cpp/htmlparser/testdata/tree-construction/menuitem-element.dat

#data
<menuitem>
#errors
10: Start tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
#document
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
</menuitem>
#errors
11: End tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
11: Stray end tag “menuitem”.
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body><menuitem>A
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"

#data
<!DOCTYPE html><body><menuitem>A<menuitem>B
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|     <menuitem>
|       "B"

#data
<!DOCTYPE html><body><menuitem>A<menu>B</menu>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|     <menu>
|       "B"

#data
<!DOCTYPE html><body><menuitem>A<hr>B
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|     <hr>
|     "B"

#data
<!DOCTYPE html><li><menuitem><li>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>
|       <menuitem>
|     <li>

#data
<!DOCTYPE html><menuitem><p></menuitem>x
#errors
39: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <p>
|         "x"

#data
<!DOCTYPE html><p><b></p><menuitem>
#errors
25: End tag “p” seen, but there were open elements.
21: Unclosed element “b”.
35: End of file seen and there were open elements.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <b>
|       <menuitem>

#data
<!DOCTYPE html><menuitem><asdf></menuitem>x
#errors
40: End tag “menuitem” seen, but there were open elements.
31: Unclosed element “adsf”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <adsf>
|     "x"

#data
<!DOCTYPE html></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><html></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><head></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><select><menuitem></select>
#errors
33: Stray start tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE html><option><menuitem>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <option>
|       <menuitem>

#data
<!DOCTYPE html><menuitem><option>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <option>

#data
<!DOCTYPE html><menuitem></body>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
<!DOCTYPE html><menuitem></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
<!DOCTYPE html><menuitem><p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <p>

#data
<!DOCTYPE html><menuitem><li>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <li>
